Added damage step logic; still needs testing.
Also: Refactored (attacker, target) args to use Battle case-class. Refactored damage step loop to use the same style of state machine as others (next() method) in order to pass information to states which were refactored to case classes. Removed ability of fast effect timing to check for trigger after chain rules in damage calc. Refactored MonsterControlledState so that it's a case class instead of a trait and impl.