Added setting of spells and traps.
Also:
Refactored card "owner" to "Owner" since it's a constant, and useful in pattern matching.
Changed Set action to SetCard to avoid naming collisions, which SetAsMonster now extends.
SpellOrTrap now has an actions method for setting.

Testing done:
Verified spells could be set right away.
Verified spells that are set and should be activated could be.
Verified spells that are set and should not be activated cannot be.