Major refactoring of actions and event timing.
ExistsInAChainAction has been renamed to Action.

Criteria has been modified to encapsulate its details further, now has a method to determine if the criteria is meetable and another to surface the available selection, on top of the validSelection method that was existing. This encapsulation has simplified methods which accept criteria as an argument, as those methods can now examine the citeria for those attributes they needed.

Actions are more composable now with the following methods: also, andIfYouDo and andThen. Particular actions have been moved to their own file. "parent"s of Actions have been removed, though Cause may be re-added.

ChainRules now keeps track of all the most recent events and CheckForTrigger is passed that list of events.

History has been removed for the moment and a TODO added for it. Similarly, the placeholders for undoing/redoing actions have been removed, though execute() still assures that an action isn't repeated accidentally.

Minor:
Cause trait has been created, though needs to be incorporated.
Some convenience utilities have been removed so as to minimize this commit (believe it or not).